+++ package.json
@@ -29,6 +29,7 @@
     "react-anchor-link-smooth-scroll": "^1.0.11",
     "react-collapsible": "^2.6.0",
     "react-dom": "^16.9.0",
+    "react-fade-in": "^0.1.6",
     "react-google-maps": "^9.4.5",
     "react-helmet": "^5.2.0",
     "react-loadable": "^5.5.0",
