+++ src/components/sections/Schedule.js
@@ -1,54 +1,37 @@
 import React from 'react';
 import styled from 'styled-components';
-
-import FullCalendar from '@fullcalendar/react';
-import dayGridPlugin from '@fullcalendar/daygrid';
-import googleCalendar from '@fullcalendar/google-calendar'
-import timeGridPlugin from '@fullcalendar/timegrid'
-import interactionPlugin from '@fullcalendar/interaction'
-
-import { Calendar, OptionsInput } from '@fullcalendar/core';
+import FadeIn from 'react-fade-in';
 
 import {Container, Section} from '@components/global';
 
-import '@styles/dog.css';
-import '@styles/shit.css';
-import '@styles/cat.css';
+import Loadable from 'react-loadable'
 
 const location = "https://goo.gl/maps/hL7Vs5zgmBAbn1dp8";
 const exposedapikey = "AIzaSyCb02sStRI4-i35sG2UMchOrs7pKDBrLq0";
 const events = 'j2addbd8s2an3stvbasq1hjitc@group.calendar.google.com';
 
+const Calendar = Loadable({
+    loader: () => import('@components/common/Calendar/Fcalendar'),
+    loading() {
+        return <div>Loading Calendar!</div>
+    },
+});
+
+
 class Schedule extends React.Component {
 
-  constructor(props){
-    super(props);
+  componentDidMount() {
     this.isMobile = window.innerWidth < 500;
   }
 
   render(){
     return(<Section id="schedule">
         <Container>
-            <div>
-              <FullCalendar
-                          defaultView={this.isMobile ? "timeGridDay" : "timeGridWeek"}
-                          header={{
-                              left: 'prev,next today',
-                              center: 'title',
-                              right: 'timeGridWeek,timeGridDay,listWeek'
-                          }}
-                          scrollTime='8:00:00'
-                          plugins={[dayGridPlugin, googleCalendar, timeGridPlugin, interactionPlugin]}
-                          events={events}
-                          googleCalendarApiKey={exposedapikey}
-                          weekends={false}
-                          eventRender={function(info) {
-                            var location = info.event.extendedProps.location || "area 51"
-                            var node = document.createElement("span")
-                            node.innerText = location
-                            info.el.appendChild(node)}}
-                />
-            </div>
+          <FadeIn delay='200'>
+            <Calendar isMobile = {this.isMobile}
+              api = {exposedapikey}
+              events = {events}/>
+          </FadeIn>
         </Container>
     </Section>)
   }
