+++ src/pages/office hours.js
@@ -1,22 +1,15 @@
 import React from 'react';
 
-import Loadable from 'react-loadable'
-
 import Layout from '@common/Layout';
 import Navbar from '@common/Navbar';
 import Footer from '@sections/Footer';
+import Schedule from '@sections/Schedule';
 
-const Calendar = Loadable({
-  loader: () => import('./../components/sections/Schedule'),
-  loading() {
-    return <div>Loading Calendar!</div>
-  },
-})
 
 const OfficeHoursPage = () => (
     <Layout>
         <Navbar/>
-        <Calendar />
+        <Schedule />
         <Footer/>
     </Layout>
 );
